Move isAttributeCacheable tests to rrweb-snapshot
